Split loginform en loginpage as components
